requirements.txt
Django==5.2.8
openai>=1.45.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0.post0

